feat: integrate person tracker
